requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
APScheduler==3.9.1
requests==2.27.1